refactor: separation fonctionnalité I/O
